Change modbus template options
Possible choices are now
- `rs485`
- `tcpip`

These define the modbus interfaces provided by the device that is defined. That `rs485` can then be accessed via a serial device (e.g. usb adapter) oder ethernet adapter is provided by the generated template and the CLI has to ask which of the two is actually used